OPT: collect items for csr_matrix in lists for its construction
Otherwise (original) code is demanding a lot of RAM majority of which is not
used (just 0s), and csr_matrix is constructed only later on.  Now we would collect
values and rows/cols to create csr_matrix from those.  It seems to be comparable
in performance and should be much more efficient in memory (runs on my laptop now),
although still somewhat wastefull requiring two copies (lists and then csr_matrix)
for a moment